+++ packages/msw/lib/auth/types.ts
@@ -1,5 +1,15 @@
+type LoginRequestBody = {
+  email: string;
+  password: string;
+};
+
+type AdminLoginRequestBody = {
+  username: string;
+  password: string;
+};
+
 type LoginResponse = {
   data: string; // JWT Token
 };
 
-export type { LoginResponse };
+export type { AdminLoginRequestBody, LoginRequestBody, LoginResponse };
