+++ packages/msw/lib/members/types.ts
@@ -1,3 +1,13 @@
+type UpdatePasswordBody = {
+  password: string;
+};
+
+type MeResponseBody = {
+  name: string;
+  generation: number;
+  position: string;
+};
+
 type Member = {
   id: number;
   name: string;
@@ -11,14 +21,28 @@ type Member = {
   isManager: boolean;
 };
 
-type MeResponseBody = {
-  name: string;
-  generation: number;
-  position: string;
-};
+type MemberRequestBody = Omit<Member, 'id'>;
 
 type MembersResponseBody = {
   data: Member[];
 };
 
-export type { Member, MembersResponseBody, MeResponseBody };
+// TODO: data 형식 (csv 포맷)으로 보낼지 확인하기
+type CreateMemberBulkBody = {
+  data: string;
+};
+
+type MemberStatusBody = Pick<Member, 'status'>;
+
+type MemberPositionBody = Pick<Member, 'position' | 'subPosition'>;
+
+export type {
+  CreateMemberBulkBody,
+  Member,
+  MemberPositionBody,
+  MemberRequestBody,
+  MembersResponseBody,
+  MemberStatusBody,
+  MeResponseBody,
+  UpdatePasswordBody,
+};
